requirements.txt
fastapi
uvicorn[standard]
python-multipart

pydicom
Pillow
numpy

torch
torchvision
